fix(test): updated some left sidebar test cases
